move DBCleaner in  pkg since it's for testing purpose only
